Layout
-Skapat komponenten Layout som skriver ut 3 knappar. Play, Highscore, Settings. Är tänkt som startsida
