+++ public/index.css
@@ -1,0 +1,12 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+}
+
+#root {
+    width: 100vw;
+    min-width: 250px;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+}
