+++ public/index.html
@@ -28,6 +28,8 @@
     <!-- Loads the Roboto font required by Material UI -->
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
 
+    <link rel="stylesheet" href="index.css">
+
     <title>React App</title>
   </head>
   <body>
