+++ src/components/QuestionCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {AnswerObject} from "../App"
+import {AnswerObject} from "../game"
 
 
 type props = {
